requirements/base.txt
Django==1.5.1
bpython==0.12
django-braces==1.2.2
django-model-utils==1.4.0
logutils==0.3.3
South==0.8.1
django-bootstrap3==1.1.0
django-userena==1.2.3
PIL==1.1.7
feedparser==5.1.3
django-grappelli==2.4.7

